refactor(bpa,proto): door errors carry u64 sizes
The size-carrying door errors move to the wire-size domain:
services::Error and cla::Error PayloadTooLarge / PayloadUnderrun
carry { size: u64, ... } (were usize), so a declared size beyond a
32-bit target's address space reports its true value — the
originate and ingress doors now hand parse_headers' u64 verdicts
through unchanged, and the last caller-side usize saturation dances
are gone. The in-memory accumulator errors (stream::ConcatError /
BufferError) deliberately keep usize: they measure buffered bytes,
not wire declarations, and widen at the door conversions instead.
buffer_stream tidies to match (the declared length converts once,
under its own name). The gRPC client's pre-flight and interim
buffering sites widen at the same boundary.

Also drops originate_raw's stale in-memory-bound comment — the door
passes the u64 cap straight through since the size bound moved into
parse_headers — and records the field widening as BREAKING in the
changelog alongside the one-seat size-bound entry.

+++ bpa/tests/forward.rs
@@ -599,7 +599,7 @@ async fn buffering_cla_rejects_under_delivering_stream() {
     assert!(matches!(
         err,
         cla::Error::PayloadUnderrun { size, expected }
-            if size == short.len() && expected == data.len()
+            if size == short.len() as u64 && expected == data.len() as u64
     ));
     assert!(events_rx.is_empty());
 }
